+++ examples/columbia_plume/ncom_forcing.py
@@ -4,7 +4,7 @@ Methods for reading NCOM ocean model outputs
 from thetis import *
 from atm_forcing import to_latlon, COORDSYS
 from thetis.timezone import *
-from thetis.forcing3d import *
+from thetis.forcing import *
 
 
 def test_time_search():
